moving data load to routes
